Extracting common logic related to integration testing of storage transforms.
